Update ntdll_winetest to Wine-23052006
svn path=/trunk/; revision=21992

+++ reactos/regtests/winetests/ntdll/large_int.c
@@ -391,22 +391,22 @@ static void test_RtlLargeIntegerToChar(void)
     result = pRtlLargeIntegerToChar(&value, 20, largeint2str[0].MaximumLength, NULL);
     ok(result == STATUS_INVALID_PARAMETER,
        "(test a): RtlLargeIntegerToChar(%llu, %d, %d, NULL) has result %lx, expected: %lx\n",
-       largeint2str[0].value, 20, largeint2str[0].MaximumLength, result, (DWORD) STATUS_INVALID_PARAMETER);
+       largeint2str[0].value, 20, largeint2str[0].MaximumLength, result, STATUS_INVALID_PARAMETER);
 
     result = pRtlLargeIntegerToChar(&value, 20, 0, NULL);
     ok(result == STATUS_INVALID_PARAMETER,
        "(test b): RtlLargeIntegerToChar(%llu, %d, %d, NULL) has result %lx, expected: %lx\n",
-       largeint2str[0].value, 20, largeint2str[0].MaximumLength, result, (DWORD) STATUS_INVALID_PARAMETER);
+       largeint2str[0].value, 20, largeint2str[0].MaximumLength, result, STATUS_INVALID_PARAMETER);
 
     result = pRtlLargeIntegerToChar(&value, largeint2str[0].base, 0, NULL);
     ok(result == STATUS_BUFFER_OVERFLOW,
        "(test c): RtlLargeIntegerToChar(%llu, %d, %d, NULL) has result %lx, expected: %lx\n",
-       largeint2str[0].value, largeint2str[0].base, 0, result, (DWORD) STATUS_BUFFER_OVERFLOW);
+       largeint2str[0].value, largeint2str[0].base, 0, result, STATUS_BUFFER_OVERFLOW);
 
     result = pRtlLargeIntegerToChar(&value, largeint2str[0].base, largeint2str[0].MaximumLength, NULL);
     ok(result == STATUS_ACCESS_VIOLATION,
        "(test d): RtlLargeIntegerToChar(%llu, %d, %d, NULL) has result %lx, expected: %lx\n",
-       largeint2str[0].value, largeint2str[0].base, largeint2str[0].MaximumLength, result, (DWORD) STATUS_ACCESS_VIOLATION);
+       largeint2str[0].value, largeint2str[0].base, largeint2str[0].MaximumLength, result, STATUS_ACCESS_VIOLATION);
 }
